requirements.txt
# 데이터 처리 및 시각화
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0

# 딥러닝 프레임워크
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0

# 임베딩 모델
sentence-transformers>=2.2.0

# LLM API
openai>=1.0.0

# 환경 변수 관리
python-dotenv>=1.0.0

# ChromaDB (현재 Python 3.14에서의 호환성 문제로 SimpleVectorStore 사용 중)
# chromadb>=0.4.22

# LangChain (추후 구현 가능)
# langchain>=0.1.0
# langchain-community>=0.0.20

# 선택적: 한국어 형태소 분석 (현재 미사용)
# konlpy>=0.6.0

